TEL suggestions

+++ MPAGSCipher/CaeserCipher.hpp
@@ -2,11 +2,8 @@
 #define MPAGSCIPHER_CAESERCIPHER_HPP
 
 #include <string>
-#include <vector>
-#include <iostream>
-#include <math.h>
 
 // Runs Caeser Cipher
-std::string CaeserCipher( const std::string& inputText, const int key, const bool encrypt);
+std::string CaeserCipher( const std::string& inputText, const size_t key, const bool encrypt);
 
 #endif // MPAGSCIPHER_CAESERCIPHER_HPP
